+++ framework/core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @forklaunch/core
 
+## 0.14.9
+
+### Patch Changes
+
+- slight hmac token creation signature change
+- Updated dependencies
+  - @forklaunch/validator@0.10.7
+  - @forklaunch/common@0.6.7
+
 ## 0.14.8
 
 ### Patch Changes